endrer adresse og flytter beboere til egen komponent
