+++ dev-server/mock-routes.js
@@ -182,7 +182,7 @@ app.get('/familie-ef-sak/api/vedtak/:behandlingId', (req, res) => {
     setTimeout(() => res.send(lesMockFil('vedtak.json')), delayMs);
 });
 
-app.post('/familie-ef-sak/api/sok/person/adresse', (req, res) => {
+app.post('/familie-ef-sak/api/sok/person/samme-adresse', (req, res) => {
     setTimeout(() => res.send(lesMockFil(`søke-person-resultat.json`)), delayMs);
 });
