+++ src/frontend/komponenter/Behandling/Personopplysninger/Adressehistorikk.tsx
@@ -1,21 +1,14 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import TabellOverskrift from './TabellOverskrift';
 import Bygning from '../../../ikoner/Bygning';
-import {
-    AdresseType,
-    IAdresse,
-    IBostedsadresse,
-    ISøkeresultatPerson,
-} from '../../../typer/personopplysninger';
+import { AdresseType, IAdresse } from '../../../typer/personopplysninger';
 import UIModalWrapper from '../../Felleskomponenter/Modal/UIModalWrapper';
 import { BredTd, KolonneTitler, TabellWrapper } from './TabellWrapper';
 import styled from 'styled-components';
 import { Knapp } from 'nav-frontend-knapper';
 import Lesmerpanel from 'nav-frontend-lesmerpanel';
-import { useApp } from '../../../context/AppContext';
-import { byggTomRessurs, Ressurs, RessursFeilet } from '../../../typer/ressurs';
-import DataViewer from '../../Felleskomponenter/DataViewer/DataViewer';
 import { datoErEtterDagensDato } from '../../../utils/utils';
+import Beboere from './Beboere';
 
 const StyledKnapp = styled(Knapp)`
     margin-left: 1rem;
@@ -34,11 +27,6 @@ const StyledLesmer = styled(Lesmerpanel)`
 
 const StyledTabellWrapper = styled(TabellWrapper)`
     padding-top: 0rem;
-`;
-
-const StyledModalTabellWrapper = styled(TabellWrapper)`
-    padding-top: 0rem;
-    grid-template-columns: max-content;
 `;
 
 const MAX_LENGDE_ADRESSER = 5;
@@ -124,52 +112,6 @@ const Innhold: React.FC<{ adresser: IAdresse[] }> = ({ adresser }) => {
                 </UIModalWrapper>
             )}
         </>
-    );
-};
-
-const Beboere: React.FC<{ adresse: IAdresse }> = ({ adresse }) => {
-    const { axiosRequest } = useApp();
-    const [søkResultat, settSøkResultat] = useState<Ressurs<ISøkeresultatPerson>>(byggTomRessurs());
-
-    useEffect(() => {
-        sokPerson();
-    }, []);
-
-    const sokPerson = () => {
-        axiosRequest<ISøkeresultatPerson, IBostedsadresse>({
-            method: 'POST',
-            url: `/familie-ef-sak/api/sok/person/adresse`,
-            data: adresse.bostedsadresse,
-        }).then((respons: Ressurs<ISøkeresultatPerson> | RessursFeilet) => {
-            settSøkResultat(respons);
-        });
-    };
-
-    return (
-        <DataViewer response={{ søkResultat }}>
-            {({ søkResultat }) => {
-                return (
-                    <>
-                        <StyledModalTabellWrapper>
-                            <table className="tabell">
-                                <KolonneTitler titler={['Navn', 'Fødselsnummer', 'Adresse']} />
-                                <tbody>
-                                    {søkResultat.hits.map((beboer, indeks) => {
-                                        return (
-                                            <tr key={indeks}>
-                                                <BredTd>{beboer.visningsnavn}</BredTd>
-                                                <BredTd>{beboer.personIdent}</BredTd>
-                                                <BredTd>{beboer.visningsadresse}</BredTd>
-                                            </tr>
-                                        );
-                                    })}
-                                </tbody>
-                            </table>
-                        </StyledModalTabellWrapper>
-                    </>
-                );
-            }}
-        </DataViewer>
     );
 };
